fix: all staticcall resp handeled correctly
