+++ tests/addresses.py
@@ -82,6 +82,7 @@ sushiswapLpTokenAddress = "0x397FF1542f962076d0BFE58eA045FfA2d347ACa0"  # USDC/W
 threeCrvAddress = "0x6c3F90f043a72FA612cbac8115EE7e52BDe6E490"
 threeCrvPoolAddress = "0xbEbc44782C7dB0a1A60Cb6fe97d0b483032FF1C7"
 triCryptoAddress = "0xc4AD29ba4B3c580e6D59105FFf484999997675Ff"
+triCryptoPoolAddress = "0xD51a44d3FaE010294C616388b506AcdA1bfAAE46"
 uniswapLpTokenAddress = "0xB4e16d0168e52d35CaCD2c6185b44281Ec28C9Dc"  # USDC/WETH
 usdc = "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48"
 usdcAddress = "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48"
